make result page

+++ router/order.js
@@ -16,8 +16,11 @@ var getTime = function() {
         return day + '-' + hour + '-' + min + '-' + sec;
 };
 
-vsprintf = require('sprintf-js').vsprintf;
-replaceAll = require('replaceall');
+var vsprintf = require('sprintf-js').vsprintf;
+var replaceAll = require('replaceall');
+
+var price_strawberry = 2500;
+var price_banana = 2000;
 
 module.exports = {
         postAdd : function(req, res) // 새로운 주문
@@ -27,36 +30,37 @@ module.exports = {
                 var n_banana = parseInt(req.params.n_banana);
                 var time = getTime();
 
-                var price_strawberry = 2500;
-                var price_banana = 2000;
+                if(Number.isInteger(bell) && Number.isInteger(n_strawberry) && Number.isInteger(n_banana)) {
+                        var price = price_strawberry * n_strawberry + price_banana * n_banana;
 
-                var price = price_strawberry * n_strawberry + price_banana * n_banana;
+                        pool = require('./pool');
+                        pool.getConnection(function(err, conn) {
+                                if(err) throw err;
 
-                pool = require('./pool');
-                pool.getConnection(function(err, conn) {
-                        if(err) throw err;
-
-                        conn.query("select * from queue where bell=" + bell + ";", function(error, results, fields) {
-                                if(results.length > 0) {
-                                        res.send("overwrite bell");
-                                        conn.release();
-                                }
-                                else {
-                                        // log에 add 정보 추가
-                                        var params = vsprintf("%d$%d$%d", [bell, n_strawberry, n_banana]);
-                                        var query = vsprintf("insert into log(type, parameter, time) values('add', %s, %s);", [pool.escape(params), pool.escape(time)]);
-                                        conn.query(query, function(error, results, fields) { });
-
-                                        // queue에 new crepe order 추가
-                                        query = vsprintf("insert into queue(bell, n_strawberry, n_banana, price, time) values(%d, %d, %d, %d, %s);", [pool.escape(bell), pool.escape(n_strawberry), pool.escape(n_banana), pool.escape(price), pool.escape(time)]);
-                                        conn.query(query, function(error, results, fields) {
-                                                res.send("complete");
-
+                                conn.query("select * from queue where bell=" + bell + ";", function(error, results, fields) {
+                                        if(results.length > 0) {
+                                                res.send("overwrite bell");
                                                 conn.release();
-                                        });
-                                }
+                                        }
+                                        else {
+                                                // log에 add 정보 추가
+                                                var params = vsprintf("%d$%d$%d", [bell, n_strawberry, n_banana]);
+                                                var query = vsprintf("insert into log(type, parameter, time) values('add', %s, %s);", [pool.escape(params), pool.escape(time)]);
+                                                conn.query(query, function(error, results, fields) { });
+
+                                                // queue에 new crepe order 추가
+                                                query = vsprintf("insert into queue(bell, n_strawberry, n_banana, price, time) values(%d, %d, %d, %d, %s);", [pool.escape(bell), pool.escape(n_strawberry), pool.escape(n_banana), pool.escape(price), pool.escape(time)]);
+                                                conn.query(query, function(error, results, fields) {
+                                                        res.send("complete");
+
+                                                        conn.release();
+                                                });
+                                        }
+                                });
                         });
-                });
+                } else {
+                        res.send("not integer");
+                }
         },
         getGet : function(req, res) // 현재 주문정보 Get
         {
@@ -100,41 +104,61 @@ module.exports = {
 
                         var id = parseInt(req.params.id);
 
-                        var completeOrder = {id : 0, bell : 0, n_strawberry : 0, n_banana : 0, price : 0, time : ''};
-                        var query = vsprintf("select id, bell, n_strawberry, n_banana, price, time from queue where id=%d;", [pool.escape(id)]);
-                        conn.query(query, function(error, results, fields) { // queue에 complete하려는 order가 존재한다면
-                                if(error) throw error;
+                        if(Number.isInteger(id)) {
+                                var completeOrder = {id : 0, bell : 0, n_strawberry : 0, n_banana : 0, price : 0, time : ''};
+                                var query = vsprintf("select id, bell, n_strawberry, n_banana, price, time from queue where id=%d;", [pool.escape(id)]);
+                                conn.query(query, function(error, results, fields) { // queue에 complete하려는 order가 존재한다면
+                                        if(error) throw error;
 
-                                if(results.length > 0)
-                                {
-                                        var rv = results[0];
-                                        completeOrder.id = rv.id;
-                                        completeOrder.bell = rv.bell;
-                                        completeOrder.n_strawberry = rv.n_strawberry;
-                                        completeOrder.n_banana = rv.n_banana;
-                                        completeOrder.price = rv.price;
-                                        completeOrder.time = rv.time;
+                                        if(results.length > 0)
+                                        {
+                                                var rv = results[0];
+                                                completeOrder.id = rv.id;
+                                                completeOrder.bell = rv.bell;
+                                                completeOrder.n_strawberry = rv.n_strawberry;
+                                                completeOrder.n_banana = rv.n_banana;
+                                                completeOrder.price = rv.price;
+                                                completeOrder.time = rv.time;
 
-                                        // Complete Task를 queue table에서 삭제
-                                        query = vsprintf("delete from queue where id=%d;", [completeOrder.id]);
-                                        conn.query(query, function(error, results, fields) { });
+                                                // Complete Task를 queue table에서 삭제
+                                                query = vsprintf("delete from queue where id=%d;", [completeOrder.id]);
+                                                conn.query(query, function(error, results, fields) {
+                                                        // Complete Task를 complete table에 추가
+                                                        query = vsprintf("insert into complete(id, n_strawberry, n_banana, price, time) values(%d, %d, %d, %d, %s);", [completeOrder.id, completeOrder.n_strawberry, completeOrder.n_banana, completeOrder.price, pool.escape(completeOrder.time)]);
+                                                        conn.query(query, function(error, results, fields) {
+                                                                if(error) throw error;
 
-                                        // Complete Task를 complete table에 추가
-                                        query = vsprintf("insert into complete(id, n_strawberry, n_banana, price, time) values(%d, %d, %d, %d, %s);", [completeOrder.id, completeOrder.n_strawberry, completeOrder.n_banana, completeOrder.price, pool.escape(completeOrder.time)]);
-                                        conn.query(query, function(error, results, fields) {
-                                                if(error) throw error;
-                                        });
+                                                                // log에 추가
+                                                                var now = getTime();
+                                                                var params = vsprintf("%d$%d$%s", [completeOrder.id, completeOrder.bell, completeOrder.time]);
+                                                                query = vsprintf("insert into log(type, parameter, time) values('complete', \'%s\', \'%s\');", [params, now]);
+                                                                conn.query(query, function(error, results, fields) {
+                                                                        if(error) throw error;
 
-                                        var now = getTime();
-                                        var params = vsprintf("%d$%d$%s", [completeOrder.id, completeOrder.bell, completeOrder.time]);
-                                        query = vsprintf("insert into log(type, parameter, time) values('complete', %s, %s);", [params, now]);
-                                        conn.query(query, function(error, results, fields) { });
+                                                                        // result 업데이트
+                                                                        query = "select * from result where id=1;";
+                                                                        conn.query(query, function(error, results, fields) {
+                                                                                if(error) throw error;
 
-                                        res.send({success : "Complete Successfully", status : 200});
-                                }
+                                                                                var rv = results[0];
+                                                                                var n_strawberry = rv.n_strawberry + completeOrder.n_strawberry;
+                                                                                var n_banana = rv.n_banana + completeOrder.n_banana;
 
-                                conn.release();
-                        });
+                                                                                query = vsprintf("update result set n_strawberry=%d, n_banana=%d where id=1;", [n_strawberry, n_banana]);
+                                                                                conn.query(query, function(error, results, fields) {
+                                                                                        if(error) throw error;
+                                                                                });
+                                                                        });
+
+                                                                        res.send({success : "Complete Successfully", status : 200});
+                                                                });
+                                                        });
+                                                });
+                                        }
+
+                                        conn.release();
+                                });
+                        }
                 });
         },
         deleteOrder : function(req, res) {
@@ -144,10 +168,34 @@ module.exports = {
 
                         // pool.escapes는 ''가 붙어나오므로 제거해준다.
                         var id = parseInt(replaceAll("\'", "", pool.escape(req.params.id)));
-                        var query = "delete from queue where id=" + id.toString() + ";";
+
+                        if(Number.isInteger(id)) {
+                                var query = "delete from queue where id=" + id.toString() + ";";
+                                conn.query(query, function(error, results, fields) {
+                                        if(error) throw error;
+                                        res.send({success : "Delete Successfully", status : 200});
+                                        conn.release();
+                                });
+                        }
+                });
+        },
+        getResult : function(req, res) {
+                pool = require('./pool');
+                pool.getConnection(function(err, conn) {
+                        if(err) throw err;
+
+                        var query = "select * from result where id=1;";
                         conn.query(query, function(error, results, fields) {
                                 if(error) throw error;
-                                res.send({success : "Delete Successfully", status : 200});
+                                var rv = results[0];
+
+                                var n_strawberry = rv.n_strawberry;
+                                var n_banana = rv.n_banana;
+                                var price = rv.price;
+
+                                rv = {n_strawberry : n_strawberry, n_banana : n_banana, price : price};
+                                res.send(rv);
+
                                 conn.release();
                         });
                 });
